streamlit_app/requirements.txt
streamlit
requests
pillow 